Add crictical fixes  from ambush
